+++ changelog.md
@@ -11,3 +11,4 @@
 * `npm run build` command
 * Do not run jshint on built tests
 * Build loads newest co and creates tests files for co tests
+* Build creates README.md